Fix inconsistent template output and update tests
There seems to be some magic happening in the templates, resulting in
lists being single or double quoted and filters ignored depending on
the versions of Python, Ansible and/or other dependencies used.

My observations are that following Jinja2 expression, when rendered via
a macro invocation:

{{ value | to_json }}

may output one of the following:

['foo','bar', 'baz']
["foo","bar","baz"]

If extra characters are added before or after the template statement the
output is always rendered correctly (e.g. the second line above). The
cause of this behaviour is not clear.

Refactoring the `telegraf_format_value` macro to print both the option
name and the value works around this issue and ensures the output is
consistent across all tested platforms.